backend/pyproject.toml
[project]
name = "ffthwatcher-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.15",
    "django-filter>=24.0",
    "psycopg[binary]>=3.2",
]

[tool.uv]
dev-dependencies = []
